// redditclone-frontend/client-codegen-api/models/post-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reddit Clone API
 * Clone Reddit With Spring Boot.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PostRequest
 */
export interface PostRequest {
    /**
     * 
     * @type {number}
     * @memberof PostRequest
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    subredditName?: string;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    description?: string;
}